refactor code to reuse same component
